Move Bundle and Console server editor parts to Eclipse Libra.
The Bundle Information, Bundle Dependency Graph and Server Console
server editor parts are moved to Eclipse Libra in a separate commit in
the Libra Git repository. This commit adapts the Virgo IDE to this
change. 

1. All moved Java classes are deleted from Virgo IDE. 
2. The plugin.xml in the UI plug-in now refers to the editor parts from
the Libra plugins. 
3. Bundle model classes now implement the respective interfaces from the
Libra bundle model. This is needed for better level of abstraction
between the editor and the data provider, so the editors can be adopted
for more servers. 
4. The editor parts now tries to adapt the IServer object to
IOSGiFrameworkAdmin and IOSGiFrameworkConsole interfaces for making the
communication with the server-side. The easiest way to make the Virgo
Server adaptable to these interfaces is to implement them in the
ServerBehavior object.